Added "Ignore" Functionality after 5 retries
This creates a "Ignore" list in your trakt.tv profile.
If the scrapers cant find a torrent/filehoster upload for a movie/show, it is added to the "Ignore" List. Items in the ignore list are ignored, even if they are in the watchlist and are not collected.
Once you remove the item from the "Ignore" Playlist, the script continues to scrape for it.